+++ Commands/Model/Teams/TeamChannelType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SharePointPnP.PowerShell.Commands.Model.Teams
+{
+    public enum TeamChannelType
+    {
+        Public,
+        Private
+    }
+}