+++ Commands/Properties/Resources.Designer.cs
@@ -79,6 +79,7 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         }
         
         /// <summary>
+
         ///   Looks up a localized string similar to The certificate located at &apos;{0}&apos; holds no private key. Ensure you provide the private key certificate, typically the .pfx file..
         /// </summary>
         internal static string CertificateAtPathHasNoPrivateKey {
@@ -86,8 +87,16 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
                 return ResourceManager.GetString("CertificateAtPathHasNoPrivateKey", resourceCulture);
             }
         }
+        ///   Looks up a localized string similar to The provided certificate with the thumbprint &apos;{0}&apos; does not have a private key which is required for a connection to be established. Ensure you have imported the certificate containing the private key, typically the .pfx file, into the Windows Certificate Store..
+        /// </summary>
+        internal static string CertificateWithThumbprintDoesNotHavePrivateKey {
+            get {
+                return ResourceManager.GetString("CertificateWithThumbprintDoesNotHavePrivateKey", resourceCulture);
+            }
+        }
         
         /// <summary>
+
         ///   Looks up a localized string similar to No private key certificate has been found at &apos;{0}&apos;.
         /// </summary>
         internal static string CertificateNotFoundAtPath {